requirements.txt
# Core
duckdb>=0.9.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# Data generation
faker>=20.0.0

# Entity resolution
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
recordlinkage>=0.15
jellyfish>=1.0.0
networkx>=3.0

# ML / Analytics
scikit-learn>=1.3.0
lifetimes>=0.11.0

# Dashboard
streamlit>=1.28.0
plotly>=5.18.0

# Dev / Testing
pytest>=7.0
pytest-cov>=4.0
ruff>=0.1
pre-commit>=3.0
